CR fixes. Add explicit serialization-json dependency instead relying on the transitive one.
