+++ gradle.properties
@@ -24,7 +24,7 @@ kotlin.code.style=official
 GROUP=com.segment.analytics.kotlin.destinations
 
 VERSION_CODE=1
-VERSION_NAME=1.0.0
+VERSION_NAME=3.0.3
 
 POM_ARTIFACT_ID=<destination>
 POM_NAME=<destination>
