+++ lib/src/main/AndroidManifest.xml
